backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-django==4.7.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
